LayoutTests/imported/w3c/web-platform-tests/css/css-logical/animation-004-expected.txt


PASS Logical properties can be transitioned
PASS Logical properties in transitions respect the writing-mode
PASS Logical properties in transitions respect the direction
PASS Declaration order is respected within declaration blocks
PASS Logical properties are able to override physical properties in declaration blocks
PASS Declaration order is respected amongst logical properties within declaration blocks
PASS Physical properties and logical properties can be mixed
PASS Declaration order is respected on each keyframe individually
FAIL Transitions update when the writing-mode is changed assert_equals: expected "0px" but got "50px"
PASS Filling transitions update when the writing-mode is changed
FAIL The number of interpolating properties can be increased when the writing-mode is changed assert_equals: expected "0px" but got "50px"
FAIL The number of interpolating properties can be decreased when the writing-mode is changed assert_equals: expected "300px" but got "275px"
FAIL Transitions update when the writing-mode is changed through a CSS variable assert_equals: expected "50px" but got "0px"
FAIL Transitions update when the direction is changed assert_equals: expected "0px" but got "50px"
FAIL Transitions from logical to physical update when the direction is changed assert_equals: expected "0px" but got "50px"
FAIL Transitions from physical to logical update when the direction is changed assert_equals: expected "0px" but got "50px"
